requirements.txt
torch==1.4.0
torchvision==0.5.0
opencv-python
opencv-contrib-python
pillow
numpy
matplotlib
tqdm